refactor(rpc): isolate native multisig simulation
