+++ crates/alloy/src/rpc/request.rs
@@ -4,128 +4,18 @@ use alloy_eips::Typed2718;
 use alloy_primitives::{Address, Bytes, U256};
 use alloy_provider::Provider;
 use alloy_rpc_types_eth::{Transaction, TransactionRequest, TransactionTrait};
-use core::{fmt, marker::PhantomData, num::NonZeroU64};
-use serde::{
-    Deserialize, Deserializer, Serialize,
-    de::{Error as _, SeqAccess, Visitor},
-};
+use core::num::NonZeroU64;
+use serde::{Deserialize, Serialize};
 use tempo_primitives::{
     AASigned, SignatureType, TempoTransaction, TempoTxEnvelope,
     transaction::{
-        Call, MAX_MULTISIG_SIGNATURES, MultisigConfig, MultisigSignature, SignedKeyAuthorization,
-        TempoSignedAuthorization, TempoTypedTransaction,
-        key_authorization::serde_nonzero_quantity_opt,
+        Call, MultisigSignature, SignedKeyAuthorization, TempoSignedAuthorization,
+        TempoTypedTransaction, key_authorization::serde_nonzero_quantity_opt,
     },
 };
 
+use super::native_multisig::MultisigSimulationWitness;
 use crate::TempoNetwork;
-
-/// Native multisig witness used only to construct an RPC simulation signature.
-#[derive(Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct MultisigSimulationWitness {
-    /// Account authorized by this witness.
-    pub account: Address,
-    /// Complete applicable configuration.
-    pub config: MultisigConfig,
-    /// Owner approvals to model.
-    #[serde(deserialize_with = "deserialize_multisig_simulation_approvals")]
-    pub approvals: Vec<MultisigSimulationApproval>,
-}
-
-/// Native multisig owner approval used for RPC simulation.
-#[derive(Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize)]
-#[serde(tag = "type", rename_all = "camelCase")]
-pub enum MultisigSimulationApproval {
-    /// Primitive owner approval.
-    Primitive {
-        /// Configured owner address.
-        owner: Address,
-        /// Signature type to model. Omission uses a maximum-size WebAuthn signature.
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        key_type: Option<SignatureType>,
-        /// Optional signature-specific gas-estimation data.
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        key_data: Option<Bytes>,
-    },
-    /// Nested native multisig owner signature.
-    Multisig {
-        /// Depth-2 multisig witness. Its approvals must all be primitive.
-        witness: MultisigSimulationNestedWitness,
-    },
-}
-
-/// Depth-2 native multisig witness used for RPC simulation.
-#[derive(Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct MultisigSimulationNestedWitness {
-    /// Nested multisig account.
-    pub account: Address,
-    /// Complete applicable configuration.
-    pub config: MultisigConfig,
-    /// Primitive owner approvals to model.
-    #[serde(deserialize_with = "deserialize_multisig_simulation_approvals")]
-    pub approvals: Vec<MultisigSimulationPrimitiveApproval>,
-}
-
-/// Primitive approval in a depth-2 multisig simulation witness.
-#[derive(Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct MultisigSimulationPrimitiveApproval {
-    /// Configured owner address.
-    pub owner: Address,
-    /// Signature type to model. Omission uses a maximum-size WebAuthn signature.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub key_type: Option<SignatureType>,
-    /// Optional signature-specific gas-estimation data.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub key_data: Option<Bytes>,
-}
-
-fn deserialize_multisig_simulation_approvals<'de, D, T>(deserializer: D) -> Result<Vec<T>, D::Error>
-where
-    D: Deserializer<'de>,
-    T: Deserialize<'de>,
-{
-    struct ApprovalsVisitor<T>(PhantomData<T>);
-
-    impl<'de, T: Deserialize<'de>> Visitor<'de> for ApprovalsVisitor<T> {
-        type Value = Vec<T>;
-
-        fn expecting(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
-            write!(
-                formatter,
-                "at most {MAX_MULTISIG_SIGNATURES} multisig simulation approvals"
-            )
-        }
-
-        fn visit_seq<A>(self, mut seq: A) -> Result<Self::Value, A::Error>
-        where
-            A: SeqAccess<'de>,
-        {
-            if seq
-                .size_hint()
-                .is_some_and(|size| size > MAX_MULTISIG_SIGNATURES)
-            {
-                return Err(A::Error::custom("too many multisig simulation approvals"));
-            }
-
-            let mut approvals = Vec::new();
-            while approvals.len() < MAX_MULTISIG_SIGNATURES {
-                let Some(approval) = seq.next_element()? else {
-                    return Ok(approvals);
-                };
-                approvals.push(approval);
-            }
-            if seq.next_element::<serde::de::IgnoredAny>()?.is_some() {
-                return Err(A::Error::custom("too many multisig simulation approvals"));
-            }
-            Ok(approvals)
-        }
-    }
-
-    deserializer.deserialize_seq(ApprovalsVisitor(PhantomData))
-}
 
 /// An Ethereum [`TransactionRequest`] extended with Tempo-specific fields.
 #[derive(
@@ -692,9 +582,9 @@ impl<P: Provider<TempoNetwork>, D: CallDecoder> TempoCallBuilderExt
 #[cfg(test)]
 mod tests {
     use super::*;
-    use alloy_primitives::{B256, Bytes, Signature, address};
+    use alloy_primitives::{Bytes, Signature, address};
     use tempo_primitives::transaction::{
-        Call, KeyAuthorization, MultisigOwner, PrimitiveSignature, TEMPO_EXPIRING_NONCE_KEY,
+        Call, KeyAuthorization, PrimitiveSignature, TEMPO_EXPIRING_NONCE_KEY,
     };
 
     fn nz(value: u64) -> NonZeroU64 {
@@ -762,37 +652,6 @@ mod tests {
         let err = serde_json::from_str::<TempoTransactionRequest>(r#"{"validAfter":"0x0"}"#)
             .expect_err("zero valid_after must be rejected during deserialization");
         assert!(err.to_string().contains("expected non-zero quantity"));
-    }
-
-    #[test]
-    fn test_deserialize_rejects_too_many_multisig_approvals() {
-        let owner = address!("0x1111111111111111111111111111111111111111");
-        let approval = MultisigSimulationApproval::Primitive {
-            owner,
-            key_type: Some(SignatureType::Secp256k1),
-            key_data: None,
-        };
-        let request = TempoTransactionRequest {
-            multisig_witness: Some(MultisigSimulationWitness {
-                account: address!("0x2222222222222222222222222222222222222222"),
-                config: MultisigConfig {
-                    salt: B256::ZERO,
-                    version: 1,
-                    threshold: 1,
-                    owners: vec![MultisigOwner { owner, weight: 1 }],
-                },
-                approvals: vec![approval; MAX_MULTISIG_SIGNATURES + 1],
-            }),
-            ..Default::default()
-        };
-        let encoded = serde_json::to_value(request).unwrap();
-
-        let error = serde_json::from_value::<TempoTransactionRequest>(encoded).unwrap_err();
-        assert!(
-            error
-                .to_string()
-                .contains("too many multisig simulation approvals")
-        );
     }
 
     #[test]
